+++ packages/@coorpacademy-components/storybook/components.js
@@ -494,7 +494,7 @@ import SettingsFixtureFreemium from '../src/organism/settings/test/fixtures/free
 import SettingsFixturePremium from '../src/organism/settings/test/fixtures/premium';
 import SettingsFixturePromoCode from '../src/organism/settings/test/fixtures/promo-code';
 import SetupHeaderFixtureDefault from '../src/organism/setup-header/test/fixtures/default';
-import SetupHeaderFixtureFromDashboards from '../src/organism/setup-header/test/fixtures/fromDashboards';
+import SetupHeaderFixtureFromDashboards from '../src/organism/setup-header/test/fixtures/from-dashboards';
 import SidebarFixtureAnalytics from '../src/organism/sidebar/test/fixtures/analytics';
 import SidebarFixtureCustom from '../src/organism/sidebar/test/fixtures/custom';
 import SidebarFixtureDashboards from '../src/organism/sidebar/test/fixtures/dashboards';
